Add files via upload
The Jupyter notebook is structured to demonstrate the use of the K-means clustering algorithm for data classification. It includes:

Data Preparation: Data is loaded and preprocessed, with necessary transformations applied to prepare it for clustering.
K-means Clustering: The K-means algorithm is applied to the data to identify clusters. Parameters such as the number of clusters are specified, and the algorithm is executed to group the data.
Visualization: The results of the clustering are visualized using scatter plots to compare the clusters formed by K-means with the actual ground truth classifications.
Evaluation: Finally, the clustering results are evaluated using metrics such as precision, recall, and F1-score to measure the effectiveness of the clustering against actual classifications.